requirements.txt
torch>=2.1.0
diffusers>=0.35.0
transformers>=4.36.0
accelerate
sentencepiece
protobuf
pillow
numpy
gradio
rembg[gpu]
onnxruntime-gpu
imageio[ffmpeg]
